% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTP.R
\name{report}
\alias{report}
\title{This function is create to make easy to people get the report of some variables of the Survey Data}
\usage{
report(y)
}
\arguments{
\item{y}{The name of the column in the Survey Data. See the metadata for the description (data_survey folder).}
}
\value{
The distribution of the respondents in each asnwer.
}
\description{
In the Paper (put the link) Section 4, we summarize some characteristics of our survey data.
This function can be use to get a quick table of the number os respondents in each asnwer of the questionnarie. With the help of metadata, just enter the name of the desired variable and get the table.
}
\examples{
report("DSE03") - returns the how much respondents we have by age group
report("HAB01") - returns the number of respondents by travel frequency - Tabel 5 in the Paper
report ("DSE07") - returns the number of respondents by family income - Table 4 in the Paper
}
